Expose `list()` endpoint under `Subscriptions` API
