+++ src/endpoints/NewsroomSubscriptions/Client.ts
@@ -1,9 +1,8 @@
 import type { DeferredJobsApiClient } from '../../api';
 import { routing } from '../../routing';
 import type { Newsroom } from '../../types';
-import { SortOrder } from '../../types';
 
-import type { CreateRequest, ListOptions, ListResponse } from './types';
+import type { CreateRequest } from './types';
 
 type NewsroomId = Newsroom['uuid'] | Newsroom['id'];
 
@@ -21,21 +20,6 @@ export class Client {
         const url = routing.newsroomSubscriptionsUrl.replace(':newsroom_id', String(newsroomId));
         return this.apiClient.post(url, {
             payload,
-        });
-    }
-
-    public async list(
-        newsroomId: NewsroomId,
-        { limit, offset, search, sortOrder }: ListOptions = {},
-    ): Promise<ListResponse> {
-        const url = routing.newsroomSubscriptionsUrl.replace(':newsroom_id', String(newsroomId));
-        return this.apiClient.get<ListResponse>(url, {
-            query: {
-                limit,
-                offset,
-                search,
-                sort: SortOrder.stringify(sortOrder),
-            },
         });
     }
 }
